refactor: zustand user store 분리 및 user profile API 수정
