+++ just-swim-app/src/app/(beforeLogin)/signup/type/page.tsx
@@ -5,93 +5,12 @@ import { useRouter, useSearchParams } from 'next/navigation';
 import { useLayoutEffect, useState } from 'react';
 import { getTokenInCookies, setTokenInCookies } from '@/(beforeLogin)/_utils';
 import { HTTP_STATUS, TEXT, USER_TYPE } from '@data';
-import { Provider, UserEntity, UserType } from '@types';
+import { UserType } from '@types';
 
-import { create } from 'zustand';
-import { createJSONStorage, persist } from 'zustand/middleware';
 import { usePostSetUserType } from '@/(beforeLogin)/_utils/server/usePostSetUserType';
 import { getMyProfile } from '@/_apis/users.ts';
 import { ROUTES } from '@/_data/routes';
-
-export type User = {
-  token: string;
-  profile: Partial<UserEntity>;
-};
-
-// TODO: store 로 따로 만들어야 한다.
-type UserStoreType = {
-  user: Record<string, Partial<Omit<User, 'token'>>>;
-  getToken: () => string;
-  getUser: () => Record<string, Partial<Omit<User, 'token'>>>;
-  getProvider: (token: Provider) => string;
-  getUserEmail: (token: string) => string;
-  getUserName: (token: string) => string;
-  getUserType: (token: string) => UserType;
-  getUserImage: (token: string) => string;
-  setAddUserToken: (token: string) => void;
-  setAddUserProfile: ({ token, profile }: User) => void;
-};
-
-export const useUserStore = create(
-  persist<UserStoreType>(
-    (set, get) => ({
-      user: {},
-      getToken: () => {
-        return Object.keys(get().user)[0];
-      },
-      getUser: () => get().user,
-      getProvider: (token: Provider) => {
-        return get().user[token]?.profile?.provider || '';
-      },
-      getUserEmail: (token: string) => {
-        return get().user[token]?.profile?.email || '';
-      },
-      getUserName: (token: string) => {
-        return get().user[token]?.profile?.name || '';
-      },
-      getUserType: (token: string) => {
-        return (get().user[token]?.profile?.userType || '') as UserType;
-      },
-      getUserImage: (token: string) => {
-        return get().user[token]?.profile?.profileImage || '';
-      },
-      setAddUserToken: (token: string) => {
-        set((state: UserStoreType) => {
-          const overWriteUsers = {
-            ...state.user,
-            [token]: { ...state.user[token], profile: {} },
-          };
-
-          return {
-            ...state,
-            user: overWriteUsers,
-          };
-        });
-      },
-      setAddUserProfile: ({ token, profile }: User) => {
-        set((state: UserStoreType) => {
-          const prevUser = state.user[token] || { profile: {} };
-          const overWriteUser = {
-            ...state.user,
-            [token]: {
-              ...prevUser,
-              profile: { ...prevUser.profile, ...profile },
-            },
-          };
-
-          return {
-            ...state,
-            user: overWriteUser,
-          };
-        });
-      },
-    }),
-    {
-      name: 'user-store',
-      storage: createJSONStorage(() => localStorage),
-    },
-  ),
-);
+import { useUserStore } from '@/_store/user';
 
 // TODO: fetch 캐싱 알아보기
 export default function Type() {
@@ -111,9 +30,8 @@ export default function Type() {
         setAddUserToken(newToken);
 
         const { data } = await getMyProfile();
-        setAddUserProfile({ token: newToken, profile: data.uesrData });
-
-        if (data.uesrData.userType) {
+        setAddUserProfile({ token: newToken, profile: data.data });
+        if (data.data.userType) {
           return router.push(ROUTES.SCHEDULE.root);
         }
         setToken(newToken);
