+++ just-swim-app/src/app/_store/index.ts
@@ -1,2 +1,2 @@
-// export * from './location';
 export * from './store';
+export * from './user';
